pipeline/requirements.txt
yfinance
pandas
google-cloud-bigquery
python-dotenv
pandas-gbq
